quick fix, changed data type to long to save turn duration

+++ Qwirkle_Beta_Version/src/qwirkle/model/Turn.java
@@ -30,7 +30,7 @@ public class Turn implements List<Move>{
         return points;
     }
 
-    public double getTurnDuration() {
+    public long getTurnDuration() {
         return turnDuration;
     }
